requirements.txt
python-dotenv
groq
gTTS 
playsound==1.2.2
fastapi
python-multipart
uvicorn
jinja2
pymongo
requests
gunicorn